jobs增加错误方法

+++ queue/lib/Worker.php
@@ -31,24 +31,20 @@ class Worker {
 
 			$body = json_decode($msg->getBody(), true);
 
-			if (empty($body) OR !isset($body['job'])) {
-				$this->delete($msg);
-				$this->failed("推送队列出错,请检查!", $body);
-				return true;
-			}
+			//没有队列对象删除消息
+			if (empty($body) OR !isset($body['job'])) $this->delete($msg);
 
-			$segments = explode('@', $body['job']);
-			$action = count($segments) > 1 ? $segments[1] : 'fire';
-
+			//实例化消耗类
 			$job = $this->makeJob($body);
 
+			//有实例化对象执行方法
 			if (is_object($job)) {
-				$this->process($job, $action, $maxTries, $body, $msg, $config);
+				$this->process($job, $this->parseJob($body, 'action'), $maxTries, $body, $msg, $config);
 			} else {
 				$this->delete($msg);
-				$this->failed("当前队列类:{$body['job']}不存在,请检查!", $body);
-				return true;
 			}
+
+			//内存超出杀死进程
 			if ($this->memoryExceeded($memory)) {
 				exit();
 			}
@@ -59,10 +55,25 @@ class Worker {
 	 * 删除当前一条队列
 	 * @param AMQPMessage $msg
 	 */
-	public function delete(AMQPMessage $msg) {
+	private function delete(AMQPMessage $msg) {
 		$this->channel = $msg->delivery_info['channel'];
 		$delivery_tag = $msg->delivery_info['delivery_tag'];
 		$this->channel->basic_ack($delivery_tag);
+	}
+
+	/**
+	 * 解析队列类和消耗方法
+	 * @param array $body
+	 * @param string $type
+	 * @return array|string
+	 */
+	private function parseJob(array $body, string $type = '') {
+		$segments = explode('@', $body['job']);
+
+		//没有消耗方法默认fire
+		$jobs = ['job' => $segments[0], 'action' => count($segments) > 1 ? $segments[1] : 'fire'];
+
+		return isset($jobs[$type]) ? $jobs[$type] : $jobs;
 	}
 
 	/**
@@ -70,7 +81,7 @@ class Worker {
 	 * @param array $body
 	 * @return Job|bool
 	 */
-	public function makeJob(array $body) {
+	private function makeJob(array $body) {
 		list($job) = explode('@', $body['job']);
 
 		if (class_exists($job)) {
@@ -86,7 +97,7 @@ class Worker {
 	 * @param  int $memoryLimit
 	 * @return bool
 	 */
-	public function memoryExceeded($memoryLimit) {
+	private function memoryExceeded($memoryLimit) {
 		return (memory_get_usage() / 1024 / 1024) >= $memoryLimit;
 	}
 
@@ -100,35 +111,27 @@ class Worker {
 	 * @param array $config
 	 * @throws Throwable
 	 */
-	public function process(Job $job, string $action, $maxTries, array $body, AMQPMessage $msg, array $config) {
-		if ($maxTries > 0 && $job->attempts() > $maxTries) {
-			//执行不能超过指定次数
-		} else {
-			try {
-				if (!$job->{$action}()) {
-					$this->failed('执行失败,未返回"true"', $body);
-				}
-			} catch (PDOException $e) {
-				$job->release($config);
-				$this->failed('执行出错,数据库错误:' . json_encode(['error' => $e->getData()]), $body);
-			} catch (Exception $e) {
-				$job->release($config);
-				$this->failed('执行出错,代码错误:' . json_encode(['error' => $e->getData()]), $body);
-			} catch (Throwable $e) {
-				$job->release($config);
-				$this->failed('执行出错,异常错误:' . json_encode(['error' => $e->getMessage()]), $body);
-			}
-		}
-		$this->delete($msg);
-	}
+	private function process(Job $job, string $action, $maxTries, array $body, AMQPMessage $msg, array $config) {
 
-	/**
-	 * 记录队列错误
-	 * @param string $msg
-	 * @param array $body
-	 * @return bool
-	 */
-	public function failed(string $msg, array $body) {
-		return true;
+		//不能执行超过最大次数
+		if ($maxTries > 0 && $job->attempts() > $maxTries) {
+			$job->setRelease(true)->failed("执行不能超过{$maxTries}次", [], $body, $config);
+			$this->delete($msg);
+			return;
+		}
+
+		//执行消耗方法
+		try {
+			if (!$job->{$action}()) $job->failed('执行后未返回"true"', [], $body, $config);
+		} catch (PDOException $e) {
+			$job->failed('执行数据库操作错误', $e->getTrace(), $body, $config);
+		} catch (Exception $e) {
+			$job->failed('执行代码异常2', $e->getTrace(), $body, $config);
+		} catch (Throwable $e) {
+			$job->failed('执行代码异常1', $e->getTrace(), $body, $config);
+		}
+
+		//执行后删除
+		$this->delete($msg);
 	}
 }
